requirements.txt
asgiref==3.8.1
chardet==5.2.0
diff-match-patch==20241021
Django==5.1.1
django-debug-toolbar==5.0.1
django-filter==24.3
django-import-export==4.3.3
django-rest-framework==0.1.0
django-simple-history==3.7.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
jdcal==1.4.1
openpyexcel==2.5.14
openpyxl==3.1.5
pillow==11.1.0
reportlab==4.2.5
sqlparse==0.5.1
tablib==3.7.0
tzdata==2024.2
